+++ great_expectations/cli/validation_operator.py
@@ -6,18 +6,13 @@ import click
 
 from great_expectations import DataContext
 from great_expectations import exceptions as ge_exceptions
-
-
-from great_expectations.cli.datasource import (
-    get_batch_kwargs,
-    select_datasource,
+from great_expectations.cli import toolkit
+from great_expectations.cli.datasource import get_batch_kwargs
+from great_expectations.cli.mark import Mark as mark
+from great_expectations.cli.util import cli_message, cli_message_dict
+from great_expectations.core.usage_statistics.usage_statistics import (
+    send_usage_message,
 )
-from great_expectations.cli.util import (
-    cli_message,
-    load_expectation_suite,
-    cli_message_dict
-)
-from great_expectations.core.usage_statistics.usage_statistics import send_usage_message
 
 json_parse_exception = json.decoder.JSONDecodeError
 
@@ -27,6 +22,7 @@ except ImportError:
     # We'll redefine this error in code below to catch ProfilerError, which is caught above, so SA errors will
     # just fall through
     SQLAlchemyError = ge_exceptions.ProfilerError
+
 
 @click.group()
 def validation_operator():
@@ -41,6 +37,7 @@ def validation_operator():
     default=None,
     help="The project's great_expectations directory."
 )
+@mark.cli_as_experimental
 def validation_operator_list(directory):
     """List known Validation Operators."""
     try:
@@ -108,6 +105,12 @@ def validation_operator_list(directory):
     "-d",
     default=None,
     help="The project's great_expectations directory.",
+)
+@mark.cli_as_deprecation(
+    """<yellow>In the next major release this command will be deprecated.
+Please consider using either:
+  - `checkpoint new` if you wish to configure a new checkpoint interactively
+  - `checkpoint run` if you wish to run a saved checkpoint</yellow>"""
 )
 def validation_operator_run(name, run_id, validation_config_file, suite, directory):
     # Note though the long lines here aren't pythonic, they look best if Click does the line wraps.
@@ -186,7 +189,7 @@ Call `great_expectation suite list` command to list the expectation suites in yo
                 )
                 sys.exit(0)
 
-            suite = load_expectation_suite(context, suite)
+            suite = toolkit.load_expectation_suite(context, suite, "cli.validation_operator.run")
 
             if name is None:
                 cli_message(
@@ -220,11 +223,11 @@ Call `great_expectation validation-operator list` command to list the operators 
 
             cli_message(
             """
-Let's help you specify the batch of data your want the validation operator to validate."""
+Let us help you specify the batch of data your want the validation operator to validate."""
             )
 
             try:
-                data_source = select_datasource(context)
+                data_source = toolkit.select_datasource(context)
             except ValueError as ve:
                 cli_message("<red>{}</red>".format(ve))
                 send_usage_message(